+++ packages/side-effects/side-effect-title/src/index.ts
@@ -3,13 +3,22 @@ import { ResponseHandler, Response } from '@curi/core';
 export interface TitleOptions {
   prefix?: string;
   suffix?: string;
+  delimiter?: string;
 }
 
 function createTitleSideEffect(options?: TitleOptions): ResponseHandler {
-  const { prefix = '', suffix = '' } = options || {};
+  const { prefix = '', suffix = '', delimiter = '' } = options || {};
 
   return function(response: Response) {
-    document.title = [prefix, response.title, suffix].join(' ');
+    const parts: Array<string> = [];
+    if (prefix !== '') {
+      parts.push(prefix, delimiter);
+    }
+    parts.push(response.title);
+    if (suffix !== '') {
+      parts.push(delimiter, suffix);
+    }
+    document.title = parts.join(' ');
   };
 }