+++ packages/side-effects/side-effect-title/tests/createTitleSideEffect.spec.ts
@@ -1,35 +1,70 @@
 import 'jest';
 import createTitleSideEffect from '../src';
-import { AnyResponse } from '@curi/core';
+import { Response } from '@curi/core';
 
 describe('createTitleSideEffect', () => {
+  const fakeResponse = {
+    title: 'Test Title; Please Ignore'
+  } as Response;
+
   it('returned function sets document.title using response.title', () => {
     const sideEffect = createTitleSideEffect();
-    const fakeResponse = {
-      title: 'Test Title; Please Ignore'
-    } as AnyResponse;
-
     const queryResponse = sideEffect(fakeResponse);
     expect(document.title).toBe('Test Title; Please Ignore');
   });
 
-  it('prepends the prefix before the title', () => {
-    const sideEffect = createTitleSideEffect({ prefix: 'My Site |' });
-    const fakeResponse = {
-      title: 'Test Title; Please Ignore'
-    } as AnyResponse;
-
+  it('sets document.title to empty string if response has no title', () => {
+    const sideEffect = createTitleSideEffect();
+    const fakeResponse = {} as Response;
     const queryResponse = sideEffect(fakeResponse);
-    expect(document.title).toBe('My Site | Test Title; Please Ignore');
+    expect(document.title).toBe('');
   });
 
-  it('appends the suffix after the title', () => {
-    const sideEffect = createTitleSideEffect({ suffix: '| My Site' });
-    const fakeResponse = {
-      title: 'Test Title; Please Ignore'
-    } as AnyResponse;
+  describe('prefix', () => {
+    it('prepends the prefix before the title', () => {
+      const sideEffect = createTitleSideEffect({ prefix: 'My Site' });
+      const queryResponse = sideEffect(fakeResponse);
+      expect(document.title).toBe('My Site Test Title; Please Ignore');
+    });
+  });
 
-    const queryResponse = sideEffect(fakeResponse);
-    expect(document.title).toBe('Test Title; Please Ignore | My Site');
+  describe('suffix', () => {
+    it('appends the suffix after the title', () => {
+      const sideEffect = createTitleSideEffect({ suffix: 'My Site' });
+      const queryResponse = sideEffect(fakeResponse);
+      expect(document.title).toBe('Test Title; Please Ignore My Site');
+    });
+  });
+
+  describe('delimiter', () => {
+    it('it adds delimiter between prefix, response.title, and suffix', () => {
+      const sideEffect = createTitleSideEffect({
+        prefix: 'Prefix',
+        suffix: 'Suffix',
+        delimiter: '|'
+      });
+      const queryResponse = sideEffect(fakeResponse);
+      expect(document.title).toBe(
+        'Prefix | Test Title; Please Ignore | Suffix'
+      );
+    });
+
+    it('does not prepend delimiter when there is no prefix', () => {
+      const sideEffect = createTitleSideEffect({
+        suffix: 'Suffix',
+        delimiter: '|'
+      });
+      const queryResponse = sideEffect(fakeResponse);
+      expect(document.title).toBe('Test Title; Please Ignore | Suffix');
+    });
+
+    it('does not append delimiter when there is no suffix', () => {
+      const sideEffect = createTitleSideEffect({
+        prefix: 'Prefix',
+        delimiter: '|'
+      });
+      const queryResponse = sideEffect(fakeResponse);
+      expect(document.title).toBe('Prefix | Test Title; Please Ignore');
+    });
   });
 });